+++ src/components/validations/leadStatuses.js
@@ -1,0 +1,16 @@
+import {required, integer} from 'vuelidate/lib/validators'
+import {alphaNumSpaceDotCommaHyphenApostrophe} from '../../shared/validation-rules'
+
+export const validations = {
+  name: {
+    required,
+    alphaNumSpaceDotCommaHyphenApostrophe
+  },
+  organizationId: {
+    required,
+    integer
+  },
+  parentId: {
+    integer
+  }
+};